feat(dashboard): generate dashboard depending on platform
