+++ charts/chainlink-cluster/dashboard/cmd/dashboard_deploy.go
@@ -1,10 +1,9 @@
 package main
 
 import (
-	"os"
-
+	"fmt"
 	"github.com/smartcontractkit/chainlink/charts/chainlink-cluster/dashboard/dashboard"
-	"github.com/smartcontractkit/wasp"
+	"os"
 )
 
 func main() {
@@ -12,38 +11,54 @@ func main() {
 	if name == "" {
 		panic("DASHBOARD_NAME must be provided")
 	}
-	ldsn := os.Getenv("LOKI_DATA_SOURCE_NAME")
-	if ldsn == "" {
-		panic("DATA_SOURCE_NAME must be provided")
+	// Can be empty
+	lokiDataSourceName := os.Getenv("LOKI_DATA_SOURCE_NAME")
+
+	prometheusDataSourceName := os.Getenv("PROMETHEUS_DATA_SOURCE_NAME")
+	if prometheusDataSourceName == "" {
+		panic("PROMETHEUS_DATA_SOURCE_NAME must be provided")
 	}
-	os.Setenv("DATA_SOURCE_NAME", ldsn)
-	pdsn := os.Getenv("PROMETHEUS_DATA_SOURCE_NAME")
-	if ldsn == "" {
-		panic("DATA_SOURCE_NAME must be provided")
-	}
-	dbf := os.Getenv("DASHBOARD_FOLDER")
-	if dbf == "" {
-		panic("DASHBOARD_FOLDER must be provided")
-	}
+
 	grafanaURL := os.Getenv("GRAFANA_URL")
 	if grafanaURL == "" {
 		panic("GRAFANA_URL must be provided")
 	}
+
 	grafanaToken := os.Getenv("GRAFANA_TOKEN")
 	if grafanaToken == "" {
 		panic("GRAFANA_TOKEN must be provided")
 	}
-	// if you'll use this dashboard base in other projects, you can add your own opts here to extend it
-	db, err := dashboard.NewCLClusterDashboard(6, name, ldsn, pdsn, dbf, grafanaURL, grafanaToken, nil)
+
+	grafanaFolder := os.Getenv("GRAFANA_FOLDER")
+	if grafanaFolder == "" {
+		panic("GRAFANA_FOLDER must be provided")
+	}
+
+	infraPlatform := os.Getenv("INFRA_PLATFORM")
+	if infraPlatform == "" {
+		panic("INFRA_PLATFORM must be provided, can be either docker|kubernetes")
+	}
+
+	err := dashboard.NewDashboard(
+		name,
+		grafanaURL,
+		grafanaToken,
+		grafanaFolder,
+		[]string{"generated"},
+		lokiDataSourceName,
+		prometheusDataSourceName,
+		infraPlatform,
+		[]string{"core"},
+		nil,
+	)
 	if err != nil {
-		panic(err)
+		fmt.Printf("Could not create dashbard: %s\n", name)
+		fmt.Printf("Error: %s\n", err)
+		os.Exit(1)
 	}
-	// here we are extending load testing dashboard with core metrics, for example
-	wdb, err := wasp.NewDashboard(nil, db.Opts())
-	if err != nil {
-		panic(err)
-	}
-	if _, err := wdb.Deploy(); err != nil {
-		panic(err)
-	}
+	fmt.Printf("Successfully deployed %s dashboard on grafana instance %s in folder %s\n",
+		name,
+		grafanaURL,
+		grafanaFolder,
+	)
 }
